+++ qoqo_for_braket/qoqo_for_braket/interface/ionq_verbatim_interface.py
@@ -60,6 +60,14 @@ def call_circuit(circuit: qoqo.Circuit) -> Circuit:
                 0.0 + qubit_phase.get(op.control(), 0.0),
                 0.0 + qubit_phase.get(op.target(), 0.0),
             )
+        elif op.hqslang() == "VariableMSXX":
+            braket_circuit.ms(
+                op.control(),
+                op.target(),
+                0.0 - qubit_phase.get(op.control(), 0.0),
+                0.0 - qubit_phase.get(op.target(), 0.0),
+                op.theta(),
+            )
         elif op.hqslang() in ALLOWED_OPERATIONS:
             pass
         else: