+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This changelog track changes to the qoqo-for-braket project starting at version 0.1.0
 
+## 0.2.6
+
+* Added VariableMSXX gate for ionq interface
+* Modified rigetti interface gates from rotatex to paulix, sqrtpaulix and invsqrtpaulix
+* Modified running with aws async program example
+
 ## 0.2.5
 
 * Moved rigetti remapping dictionary to verbatim box in rigetti interface